Don't strip slashes when outputting JSON body objects
JSON is intended to be a "safe" format, and stripping slashes here
doesn't add any security that I can see.

It does - however - break copying and pasting of these blocks. If I
search for やばい on an ElasticPress-powered site, and then copy the
query body to paste into Kibana to explore further, I would want to get

```
  "query": "\u3084\u3070\u3044",
```

rather than

```
  "query": "u3084u3070u3044",
```

This removes the `stripslashes` escaping from each of the pre-formatted
JSON blobs in the ElasticPress Query Monitor panel: "Query Body", "Query
Result", and "Query Errors".